MudExtensions Enums namespace and some component page refactoring
