+++ CodeBeam.MudExtensions/Components/DateWheelPicker/MudDateWheelPicker.razor
@@ -1,5 +1,6 @@
 ﻿@namespace MudExtensions
 @inherits MudBaseInput<DateTime?>
+@using MudExtensions.Enums
 
 <CascadingValue Name="SubscribeToParentForm" Value="false" IsFixed="true">
     <div>
