+++ CodeBeam.MudExtensions/Components/DateWheelPicker/MudDateWheelPicker.razor.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Components.Web;
 using Microsoft.Extensions.DependencyInjection;
 using MudBlazor;
 using MudBlazor.Utilities;
+using MudExtensions.Enums;
 using static MudBlazor.CategoryTypes;
 
 namespace MudExtensions
