+++ CodeBeam.MudExtensions/Components/Loading/MudLoading.razor
@@ -1,5 +1,6 @@
 ﻿@namespace MudExtensions
 @inherits MudComponentBase
+@using MudExtensions.Enums
 
 
 @if (Loading == true)
