+++ CodeBeam.MudExtensions/Components/LoadingButton/MudLoadingButton.razor
@@ -1,5 +1,6 @@
 ﻿@namespace MudExtensions
 @inherits MudBaseButton
+@using MudExtensions.Enums
 
 @if (ButtonVariant == ButtonVariant.Button)
 {
