+++ CodeBeam.MudExtensions/Components/LoadingButton/MudLoadingButton.razor.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
 using MudBlazor;
 using MudBlazor.Utilities;
+using MudExtensions.Enums;
 
 namespace MudExtensions
 {
