+++ CodeBeam.MudExtensions/Components/Page/MudPage.razor.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Components.Web;
 using MudBlazor;
 using MudBlazor.Extensions;
 using MudBlazor.Utilities;
+using MudExtensions.Enums;
 
 namespace MudExtensions
 {
