+++ CodeBeam.MudExtensions/Components/Popup/MudPopup.razor
@@ -1,5 +1,6 @@
 ﻿@namespace MudExtensions
 @inherits MudComponentBase
+@using MudExtensions.Enums
 
 @if (Open)
 {
