+++ CodeBeam.MudExtensions/Components/Popup/MudPopup.razor.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
+using MudExtensions.Enums;
 
 namespace MudExtensions
 {
