+++ ComponentViewer.Docs/Pages/Components/LoadingButtonPage.razor
@@ -1,110 +1,12 @@
 ﻿@page "/mudloadingbutton"
-
-@inject ISnackbar Snackbar
+@using ComponentViewer.Docs.Pages.Examples
 
 <ExamplePage Title="MudLoadingButton">
-    <ExampleCard Title="Usage" Description="MudLoadingButton has required parameters to achieve easy loading control.">
-        <MudGrid>
-            <MudItem xs="12" sm="8" Class="d-flex align-center flex-wrap">
-                <MudStack Spacing="8">
-                    <MudStack Row="true" Spacing="4">
-                        <MudLoadingButton @bind-Loading="_loading" Size="Size.Small" LoadingAdornment="_adornment" OnClick="ButtonClick" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                        <MudLoadingButton @bind-Loading="_loading" OnClick="ButtonClick" StartIcon="@Icons.Custom.Brands.MudBlazor" EndIcon="@Icons.Custom.Brands.MudBlazor" LoadingAdornment="_adornment" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                        <MudLoadingButton @bind-Loading="_loading" Size="Size.Large" LoadingAdornment="_adornment" OnClick="ButtonClick" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                    </MudStack>
-
-                    <MudStack Row="true" Spacing="4">
-                        <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.IconButton" Size="Size.Small" OnClick="ButtonClick" Icon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                        <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.IconButton" Size="Size.Medium" OnClick="ButtonClick" Icon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                        <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.IconButton" Size="Size.Large" OnClick="ButtonClick" Icon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                    </MudStack>
-
-                    <MudStack Row="true" Spacing="4" AlignItems="AlignItems.Center">
-                        <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.Fab" Size="Size.Small" OnClick="ButtonClick" StartIcon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                        <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.Fab" Size="Size.Medium" OnClick="ButtonClick" StartIcon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                        <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.Fab" Size="Size.Large" OnClick="ButtonClick" StartIcon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
-                    </MudStack>
-                </MudStack>
-            </MudItem>
-
-            <MudItem xs="12" sm="4">
-                <MudStack>
-                    <MudSwitch @bind-Checked="_loading" Color="Color.Primary">Loading</MudSwitch>
-                    <MudSelect @bind-Value="_adornment" Variant="Variant.Outlined" Label="Loading Adornment">
-                        <MudSelectItem Value="Adornment.Start">Start</MudSelectItem>
-                        <MudSelectItem Value="Adornment.End">End</MudSelectItem>
-                    </MudSelect>
-                </MudStack>
-            </MudItem>
-        </MudGrid>
+    <ExampleCard ExampleName="LoadingButtonExample1" Title="Usage" Description="MudLoadingButton has required parameters to achieve easy loading control.">
+        <LoadingButtonExample1 />
     </ExampleCard>
 
-    <ExampleCard Title="Custom Loading Fragment" Description="You can change loading text with LoadingContent parameter. Supports render fragment. Shows also with IconButton and Fab variants if set.">
-        <MudGrid>
-            <MudItem xs="12" sm="8" Class="d-flex gap-4 align-center flex-wrap">
-                <MudLoadingButton @bind-Loading="_loading2" LoadingAdornment="_adornment2" LoadingCircularColor="Color.Primary" OnClick="ButtonClick2" Variant="Variant.Filled" Color="Color.Primary">
-                    <LoadingContent>
-                        LOADING
-                    </LoadingContent>
-                    
-                    <ChildContent>
-                        CLICK ME
-                    </ChildContent> 
-                </MudLoadingButton>
-
-                <MudLoadingButton ButtonVariant="ButtonVariant.IconButton" @bind-Loading="_loading2" Icon="@Icons.Custom.Brands.MudBlazor" LoadingAdornment="_adornment2" LoadingCircularColor="Color.Primary" OnClick="ButtonClick2" Variant="Variant.Filled" Color="Color.Primary">
-                    <LoadingContent>
-                        LOADING
-                    </LoadingContent>
-                    
-                    <ChildContent>
-                        CLICK ME
-                    </ChildContent> 
-                </MudLoadingButton>
-
-                <MudLoadingButton ButtonVariant="ButtonVariant.Fab" @bind-Loading="_loading2" StartIcon="@Icons.Custom.Brands.MudBlazor" LoadingAdornment="_adornment2" LoadingCircularColor="Color.Primary" OnClick="ButtonClick2" Variant="Variant.Filled" Color="Color.Primary">
-                    <LoadingContent>
-                        LOADING
-                    </LoadingContent>
-                    
-                    <ChildContent>
-                        CLICK ME
-                    </ChildContent> 
-                </MudLoadingButton>
-            </MudItem>
-
-            <MudItem xs="12" sm="4">
-                <MudStack>
-                    <MudSwitch @bind-Checked="_loading2" Color="Color.Primary">Loading</MudSwitch>
-                    <MudSelect @bind-Value="_adornment2" Variant="Variant.Outlined" Label="Loading Adornment">
-                        <MudSelectItem Value="Adornment.Start">Start</MudSelectItem>
-                        <MudSelectItem Value="Adornment.End">End</MudSelectItem>
-                    </MudSelect>
-                </MudStack>
-            </MudItem>
-        </MudGrid>
+    <ExampleCard ExampleName="LoadingButtonExample2" Title="Custom Loading Fragment" Description="You can change loading text with LoadingContent parameter. Supports render fragment. Shows also with IconButton and Fab variants if set.">
+        <LoadingButtonExample2 />
     </ExampleCard>
 </ExamplePage>
-
-@code{
-    bool _loading = false;
-    Adornment _adornment = Adornment.Start;
-
-    bool _loading2 = false;
-    Adornment _adornment2 = Adornment.Start;
-
-
-    private async Task ButtonClick()
-    {
-        _loading = true;
-        await Task.Delay(2000);
-        _loading = false;
-    }
-
-    private async Task ButtonClick2()
-    {
-        _loading2 = true;
-        await Task.Delay(2000);
-        _loading2 = false;
-    }
-}
