+++ ComponentViewer.Docs/Pages/Examples/LoadingButtonExample1.razor
@@ -1,0 +1,45 @@
+﻿<MudGrid>
+    <MudItem xs="12" sm="8" Class="d-flex align-center flex-wrap">
+        <MudStack Spacing="8">
+            <MudStack Row="true" Spacing="4">
+                <MudLoadingButton @bind-Loading="_loading" Size="Size.Small" LoadingAdornment="_adornment" OnClick="ButtonClick" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+                <MudLoadingButton @bind-Loading="_loading" OnClick="ButtonClick" StartIcon="@Icons.Custom.Brands.MudBlazor" EndIcon="@Icons.Custom.Brands.MudBlazor" LoadingAdornment="_adornment" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+                <MudLoadingButton @bind-Loading="_loading" Size="Size.Large" LoadingAdornment="_adornment" OnClick="ButtonClick" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+            </MudStack>
+
+            <MudStack Row="true" Spacing="4">
+                <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.IconButton" Size="Size.Small" OnClick="ButtonClick" Icon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+                <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.IconButton" Size="Size.Medium" OnClick="ButtonClick" Icon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+                <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.IconButton" Size="Size.Large" OnClick="ButtonClick" Icon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+            </MudStack>
+
+            <MudStack Row="true" Spacing="4" AlignItems="AlignItems.Center">
+                <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.Fab" Size="Size.Small" OnClick="ButtonClick" StartIcon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+                <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.Fab" Size="Size.Medium" OnClick="ButtonClick" StartIcon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+                <MudLoadingButton @bind-Loading="_loading" ButtonVariant="ButtonVariant.Fab" Size="Size.Large" OnClick="ButtonClick" StartIcon="@Icons.Custom.Brands.MudBlazor" Variant="Variant.Filled" Color="Color.Primary">CLICK ME</MudLoadingButton>
+            </MudStack>
+        </MudStack>
+    </MudItem>
+
+    <MudItem xs="12" sm="4">
+        <MudStack>
+            <MudSwitch @bind-Checked="_loading" Color="Color.Primary">Loading</MudSwitch>
+            <MudSelect @bind-Value="_adornment" Variant="Variant.Outlined" Label="Loading Adornment">
+                <MudSelectItem Value="Adornment.Start">Start</MudSelectItem>
+                <MudSelectItem Value="Adornment.End">End</MudSelectItem>
+            </MudSelect>
+        </MudStack>
+    </MudItem>
+</MudGrid>
+
+@code {
+    bool _loading = false;
+    Adornment _adornment = Adornment.Start;
+
+    private async Task ButtonClick()
+    {
+        _loading = true;
+        await Task.Delay(2000);
+        _loading = false;
+    }
+}
