+++ ComponentViewer.Docs/Pages/Index.razor
@@ -31,6 +31,11 @@
             <ComponentCard Title="API"
                            Description="A basic API list for all components." />
         </MudItem>
+        
+        <MudItem xs="12" sm="6" md="4" lg="3" xl="2">
+            <ComponentCard Title="MudAnimate"
+                           Description="A revolutionary next generation animate component." />
+        </MudItem>
 
         <MudItem xs="12" sm="6" md="4" lg="3" xl="2">
             <ComponentCard Title="MudDateWheelPicker"
@@ -45,11 +50,6 @@
         <MudItem xs="12" sm="6" md="4" lg="3" xl="2">
             <ComponentCard Title="MudInputStyler"
                            Description="Applies CSS colors easily for mud inputs like MudTextField and MudSelect." />
-        </MudItem>
-
-        <MudItem xs="12" sm="6" md="4" lg="3" xl="2">
-            <ComponentCard Title="MudAnimate"
-                           Description="A revolutionary next generation animate component." />
         </MudItem>
 
         <MudItem xs="12" sm="6" md="4" lg="3" xl="2">
