+++ ComponentViewer.Docs/_Imports.razor
@@ -11,3 +11,4 @@
 
 @using MudBlazor
 @using MudExtensions
+@using MudExtensions.Enums
