+++ ComponentViewer.Wasm/_Imports.razor
@@ -10,3 +10,4 @@
 
 @using MudBlazor
 @using MudExtensions
+@using MudExtensions.Enums
